ticket 28: collector at 761 instruments — same cost per instrument, one socket per ~380
Owner 2026-09-12: "the whole point was a collector that can record not 10
coins but, say, 500 just as efficiently and economically".

- `lob session --all-instruments` (every linear USDT perpetual, no §2
  exclusions — this is recording, not the analysis pool)
- subscriptions are packed per socket up to Bybit's documented `args` limit
  (`MAX_ARGS_CHARS = 21_000`, quoted with URL and date); one I/O thread per
  socket, one decision thread, one mpsc channel (A3, D04); per-socket routing
  by binary search, no HashMap on the event path; `feed::Event.symbol` u16
- a socket close fans `Disconnected` out to every instrument of that socket
  (`reconnects` counts the socket once, `gaps.csv` one row per instrument);
  a market frame with an unrouted symbol is counted (`session.json.unrouted`)
- layout errors (`EmptyPool`, `PoolTooLarge`) instead of zero shards or a
  clamped index; `session.json` rewritten once after a midnight rotation of
  all instruments

Measured live, 5 min each (docs/findings/collector-2026-09-12.md, "761"):
8 → 761 instruments, sockets 1 → 2, CPU 2.0 → 24.2 % of a core (0.253 →
0.032 % per instrument), RSS 17.8 → 159.5 MiB (2 283 → 215 KiB per
instrument), 199 → 5 255 KB/min (24.9 → 6.9 per instrument), bytes/record
7.54 → 6.38, parse p99 56 → 87 µs (budget 200), queue p99 1.9 → 27.8 ms,
gaps/reconnects/resyncs/frames_failed 0; ≈ 7.75 GB/day. Named and not
fixed: queue p99 on 761 breaks the live-bot budget; RSS not on a plateau
within 5 min.

Review: one blocking finding (socket close reached one instrument) closed in
a repair round; the rest recorded as concerns.

+++ src/feed/replay.rs
@@ -24,7 +24,7 @@ use super::{Event, Feed};
 
 /// `Feed` с одного суточного файла бинлога одного инструмента пула.
 pub struct ReplayFeed<R: Read> {
-    symbol: u8,
+    symbol: u16,
     tick_e9: i64,
     step_e9: i64,
     reader: binlog::Reader<R>,
@@ -37,7 +37,7 @@ pub struct ReplayFeed<R: Read> {
 impl<R: Read> ReplayFeed<R> {
     /// `symbol` — тот же индекс пула, под которым файл писала `lob session`
     /// (см. doc модуля `feed`: тег события не имеет права быть `String`).
-    pub fn open(symbol: u8, inner: R) -> Result<Self, BinlogError> {
+    pub fn open(symbol: u16, inner: R) -> Result<Self, BinlogError> {
         let reader = binlog::Reader::open(inner)?;
         let header = reader.header();
         Ok(Self {
